Ensure reducer values remain lazy.
Fix bug where the set of values passed into a reduce function are
completely brought into memory rather than being treated as a true lazy
Iterable.

As a consequene of this fix, if there is no reducer (i.e. the outptut of
a GroupByKey is the last node in an MSCR resulting in an identity
reducer), need to ensure that the Iterable values *are* actually brought
into memory before emitting.