Refactor implementation of data I/O
Refactor how I/O is implemented to make it simpler to support different
I/O types. For example, without this refactoring, it is difficult to
implement "native" support for Sequence files.

The primary refactoring is for MapReduce jobs generated by Scoobi to
use InputFormats and OutputFormats specific to the type of I/O required.
This is in contrast to the current implementation where I/O was
hard-coded to text files with BridgeStores being a special case.
Using InputFormats and OutputFormats in their raw form allows the many
exsiting classes to be leveraged. A consequence of this change is that
InputConverters and OutputConverters are implmeneted to convert between
the key-value types of InputFormats/OutputFormats and the types of
DLists.

As a demonstration of this refactoring, also implmenent support for
reading and writing Sequence Files.