Create DLists from a Scala function.
Add a DataSource that produces a specifed number of input values based
on a Scala function that maps an index to an element value. The
DataSource requires the creation of a FunctionInputFormat that produces
InputSplits and RecordReaders encapuslating the specifie ScalaFunction.

Using the new DataSource, create DList companion object "sugar" methods
for easily forming new DLists from a list/sequence of values or by
directly specifying a Scala function.