Improve logging infromation of steps.
Add more logging for information associated with each of the 'steps'
(i.e. MSCR) executed in a Scoobi job. Further, log 'job' level
information such as the number of tasks, input sources, output targets,
etc. Modify the MR job name for Scoobi tasks to be a concatenation of
the Scoobi job id and current step id.