Implement zipWithIndex method on DList.
The zipWithIndex method for DList will have similar semantics to the
zipWithIndex method in the standard Scala collections. In particualr,
for a DList of length n, each integer in the range 0 to n - 1 will be
present exactly once. Of course, as DLists themselves are not ordered,
indices will not be ordered either.